core: Add new offload provider
Introduce a new type of provider, known as an offload
provider.  Generically, this is a provider
that only implements a specific set of functionality for
use by a main provider.  It can be considered a type of
peer provider, but one with severe restrictions.

The main target of this new provider type is a provider
that implements support for collective operations, and
more specifically a provider which supports switch-based
collectives.  This architecture will allow a provider
that supports in-network compute/collective operations
to be paired with any main provider that supports the same
transport or link protocols.  In this model, collective
operations would be routed to the function provider.

An offload provider is, generally, not usable directly by
an application, as it will place requirements on the main
provider before it may be used.  Additional details will
be in follow-up patches.  This patch merely adds the very
basic of frameworks.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>