prov/efa: remove de-allocated CUDA memory region during registration
This patch is a temporary workaround for issue:

	https://github.com/ofiwg/libfabric/issues/8081

which is current cuda memory monitor does not monitor the event of
cuda memory being freed, causing freed memory regions remain in cache.

The long term fix is to implement a cuda memory monitor that monitors
cuda memory free event. Before that is implemented, this patch
will detect such freed cuda memory region when registering a new
memory region, and remove it from cache.

This patch is to be reverted after proper cuda memory monitor was
implemented and used by efa.

Signed-off-by: Wei Zhang <wzam@amazon.com>