prov/shm: implement shared receive queue internally
This adds a shared receive queue abstraction layer as initial steps to add
full external srx support. This adds all the srx functions in shm for shm
to be either an owner or peer. Since no other providers implement shared
functionality, shm will always be both the owner and peer.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>