core/peer: Use a single FI_PEER flag
Rather than defining a peer flag for every object type that might
be shared, using a common flag.  This reduces the number of flags
that we have to consume.

We don't yet have a capabililty bit to indicate if a provider
supports peer objects.  But it's expected that any such bit would
need to indicates that a full set of peer capabilities are supported
for the provider to be usable as a peer.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>