prov/cxi: Extend test framework.
Remove obsolete test infrastructure files. All distributed testing
should be kept in the distributed directory, and should be built on the
pmi_frmwk.

Extended to support the SLURM PMI variant.

PMI does not always work properly if included through a shared library.
Code modified to ensure PMI is initialized and terminated in the main()
function and included statically.

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>