hmem/ze: Add support for host register/unregister
The functionality is provided as ZE extensions so the initialization is
slightly different from the rest.

Also fixed some function names for consistency.

Moved the definition of ze_hmem_get_ipc_handle_size() out from a
conditional to fix an undefined symbol error when drm headers are
not present on the system.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>