core/getinfo: Add ability to re-sort the fi_info list
The current sorting of fi_info is based purely on the provider
order.  However, with the enhancements made to the net provider
(soon to be tcp again), it is desirable to report that RDM EPs
can be supported directly as well as layered with rxm.  But,
since rxm has preference (in order to prioritize rxm/verbs),
rxm;tcp will also have priority over tcp, even though, tcp
by itself will perform better without the rxm layering.  (The
net/tcp provider integrates the rxm connection multiplexing
directly).

To fix this add the ability to re-order the fi_info list based
on a more complex set of attributes.  Implement this for the
net provider to move its rdm ep support ahead of rxm;net.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>