Add utility collective provider
The utility collective provider (named "coll") is an "offload" type
provider that implements collective operations over point-to-point
communication mechanism provided by the peer (main) provider. It
not only serves as an example of the peer provider API but also can
be used by another collective offload provider as the "bootstrap"
collectives needed by the HW offload engines.

It replaces the existing utility collective implementation under
"prov/util/src", which will be removed in a later commit.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>