prov/net: Fix possible crash handling FI_DISCARD
We need to allocate an internal buffer to receive a discarded
message.  Once the message has been received, we can free the
buffer.  The buffer is stored in the recv_entry iov[0] as
iov_base.  The iov_base is freed later.  However, as part of
the receive handling, the iov_base can be updated.  This is
done to handle partial receives, where we need to wait for
additional data to arrive from the peer before receiving
more data.

To fix, we need to record the address of the allocation
in a separate location, so we can free it later.  Use
the existing user_buf pointer for this, which is used
in other places to track the start of the buffer.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>