fabtests/rdm_rma_trigger: add sync to make sure counter values match
The start_rx and start_tx values of the counters are read before starting
the test to remove the need for hardcoded values in case the common code
adds more sends and receives. 
However, there is no synchronization after that read call so there is a
race condition where the rx counter could have already read one of the test
RMA calls which puts the whole test out of sync.
This patch adds that missing sync call.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>