prov/cxi: Add missing MADV_DONTFORK for EQ buffers
If the EQ buffer is allocated using mmap and is not mapped using
cxil_map, the EQ buffer can be subjected to copy-on-write if the
application does a fork. The result of this is the parent process
getting new pages leading to hangs since events are being written into
the child's memory.

To prevent this issue, call MADV_DONTFORK on the EQ buffer.

NETCASSINI-4261

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>