util/rxm_av: Add option to free resources on AV removal
When sending or receiving messages, an AV may allocate
resources underneath, such as establishing a connection.
These connections can persist even if the peer's address
is not inserted into the AV or after it's been removed.
This can cause use of resources as long as the peer is
running, but no longer actively communicating with
another peer.

Add the ability to cleanup any underlying resources when
removing an address from an AV.  This can free receive
buffers, close fd's, etc. to reduce resource utilization.

This capability is disabled by default as it may abruptly
terminate communication with a peer if there are data
transfers in progress.  To enable, a libfabric wide
environment variable must be set.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>