hmem/rocr: Support asynchronous memory copies
A key performance improvement for shm provider is to utilize asynchronous
memory copies provided by the device, if available.

This implementation targets ROCR, however, it can be expanded to handle
other devices.

ROCR adds a rocr_dev_async_copy()
This API creates a ROCR signal and triggers an asynchronous write, passing
it the signal which was created. When the write completes the signal will
be updated.

Add a freestack of signals used for the asynchronous calls
Add a freestack of stream objects
Modify the free stack to add a destroy callback to clean up the signals

The freestacks are protected by a spinlock. These are global stacks which
can be accessed concurrently if there are multiple progress threads

This async copy operation is abstracted in the hmem.c file via

static ssize_t
ofi_async_copy_hmem_iov_buf(enum fi_hmem_iface hmem_iface, uint64_t device,
			    const struct iovec *hmem_iov,
			    size_t hmem_iov_count,
			    uint64_t hmem_iov_offset, void *buf,
			    size_t size, int dir, ofi_hmem_async_even_t event)

This API resembles the none async version, however, it takes an abstract
object, event. An event is meant to track all asynchronous operations for
a single copy. For example, if we have multiple IOVs and we trigger an
async copy for each entry in the IOV, the copy would only be considered
complete when all copies are completed.

Different calls to the upper level async copy API will trigger concurrent
asynchronous memory operations.

Events are allocated and freed by the caller.

In the ROCR implementation an event object points to an array of ROCR
signal pointers. For each call to the copy operation on the same event, a
ROCR signal is added to this async copy operation. In essence if you have
an IOVEC with 4 entries, these 4 entries trigger 4 concurrent memory
operations.

Calls on different events trigger concurrent memory operations.

int ofi_async_copy_query(enum fi_hmem_iface iface,
                         ofi_hmem_async_even_t event)

The query API checks that all signals associated with an event are now
complete. If all are complete then it returns FI_SUCCESS. This allows the
operation to be completed.

The user of this infrastructure will need to implement the following work
flow:

  . Create an asnyc event
  . Trigger an Async operation
  . keep track of the async event
  . Progress this list periodically
  . On every progress operation, iterate over the list and query each
    async operation for completeness
  . Process completed operations appropriately
  . free the async event once it's completed.

The HSA API is used in this implementation because it uses the DMA engine
which is dedicated for IO operations. The HIP API uses the BLIT Kernel,
used for computation. This would cause IO to compete with compute
operations.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>