prov/shm: Optimize locking
The locking which the SHM provider uses is quiet coarse. It locks the
shared memory region lock across the entire memory operation. This would
force the processes which are sending to the same peer, to serialize their
operations.

Use the atomic queue for the command queue.
This allows posting commands without the need to acquire the region lock.
Use the region lock for protecting the inject buffer pool and SAR list.

With the introduction of this lock optimization work it is now possible for
multiple progress threads to match messages out of order. This breaks the
msg_order requirements stated here:
https://ofiwg.github.io/libfabric/main/man/fi_endpoint.3.html

To avoid this, use the ep->util_ep lock. This lock is held
before a message is dequeued off the command queue, and held until the
matching is completed. It's unlocked for all other cases.

With multiple progress threads, this lock will serialize the matching
process while at the same time allowing parallelism for the actual memory
operations.

This should be a low cost lock. It is not a shared lock, and therefore
other processes are free to post commands on the command queue, without
forcing any serialization between processes as in the case with using a
region lock.

With the new queue implementation, there is no need to keep track of the
cmd_cnt, as it doesn't add much use.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>