hmem: Define HMEM dev reg/unreg
ofi_hmem_dev_register() is used to perform additional memory
registration against a given HMEM device pointer to enable optimized
memcpy routines if supported. In addition, if the HMEM stack supports
it, a memory mapped CPU pointer to the device memory will be provided.
This can be used to do direct loads and stores instead of going to HMEM
specific copy routines.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>