prov/cxi: Integrate host_addr into IDC path
For messages which are sent via the IDC path and are not FI_INJECT, if
FI_HMEM is set, the memory is registered. Memory registration will
result in the host_addr being set in the memory descriptor. If this is
set, CXI provider can do direct loads from device memory. This can
significantly decrease latency with device memory.

NETCASSINI-4318

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>