prov/cxi: Update cxip_txc_copy_from_hmem
Any optional memory descriptor can now be provided used to optimize the
memcpy.

cxip_txc_copy_from_hmem() is no longer needed.

If the hmem_md is NULL and FI_HMEM is set for the TXC, the hmem_src is
temporarily registered. This enables the ability to do load from device
memory using a CPU pointer.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>