infra: Add atomic queue
Add atomic queue infrastructure, which does not need to be protected.

Example usage for this data structure is to build a command queue shared
between different processes. Multiple processes would post commands on a
command queue which belongs to another process. The receiving process
would read and process commands off the queue.

Usage:
  . OFI_DECLARE_ATOMIQ() to declare the atomic queue
  . Allocate shared memory for the queue or call the create method
  . call the init() method to ready the queue for usage
  . To post on the queue call tx_next() method
     . This will return a buffer of entrytype
     . Initialize the entry
  . Call tx_advance() method to post for the reader
  . To read off the queue call rx_next()
     . This will return the next available entry on the queue or
       -FI_ENOENT if there are no more entries on the queue
  . Call the rx_advance() after you've completed using the entry

Signed-off-by: Amir Shehata <shehataa@ornl.gov>