prov/net: Fetch correct ep when posting to a loopback connection
The conn_idx_map contains a list of connections corresponding to
peer addresses.  Basically, there's a connection per AV entry,
with a separate connection map per rdm ep.  When sending data,
we use the AV index into the conn_idx_map to retrieve the
connection to send to.  We also use the index when posting a
receive buffer to match with the connection that a message will
arrive on.

This works fine, except for one case -- loopback connections.
Loopback connections have 2 sockets, one used to post sends to
and the other where the data arrives.  The distinction is critical,
as the sockets will report different events (POLLOUT when ready to
send and POLLIN when data can be read).

In xnet_srx_tag(), when posting a message with source address
matching, we call xnet_get_ep() to get the message endpoint
associated with the receive's src_addr.  In the case of loopback
connections, this has the result of returning the send-side
endpoint.  In the case of an unexpected message, the data is
already queued on the receive-side endpoint.

To correct for this, we need to track if an endpoint is part of
a loopback connection.  If so, when a receive is posted, we need
to use the receive-side endpoint instead of the send-side endpoint.
This fixes a hang running certain MPI tests, such as mpich bsend1.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>