fabtests/fi_rdm_atomic: Fix hmem support
Before this patch, alloc_ep_res() would allocate buffers on the host,
but ft_reg_mr() uses opt.iface to register the buffers as cuda buffers.
This caused the EFA provider to segfault during memory registration.
This patch allocates the result/compare buffers on the device, which
allows the memory to be successfully registered.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>