prov/efa: fix a bug rxr_pkt_req_hdr_size()
rxr_pkt_req_hdr_size() calculates the exact header size of a REQ
packet.

Prior to this patch, it assumes the opt raw headersize to be
a constant: RXR_REQ_OPT_RAW_HDR_SIZE. The assumption is wrong
because this header size is a variable and is stored in as
"addr_len" in rxr_req_opt_raw_hdr.

This patch addressed the issue by:

1. merging rxr_pkt_req_hdr_size_from_entry() into
   rxr_pkt_req_hdr_size(). The new function takes rxr_pkt_entry
   as input and load correct opt raw address header size from
   pkt entry's header

2. Re implement rxr_pkt_req_max_hdr_size() such that
   it does not call rxr_pkt_req_hdr_size().

3. Re implement rxr_op_entry_max_data_capacity that it call
   rxr_pkt_max_hdr_size() instead of rxr_pkt_req_hdr_size().

Signed-off-by: Wei Zhang <wzam@amazon.com>