prov/efa: HMEM support for emulated atomics
Add emulated atomic support for HMEM to the EFA provider.
For HMEM Cuda, Atomics are only supported on fabrics that are
using CUDA Memcpy. The atomic operations uses a bounce buffer
on the remote side. The remote does a memcpy from the device
to a bounce buffer, does the atomic operation on the bounce
buffer, and then copies the result back to device memory.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>
Co-authored-by: Luke Robinson <lrbinson@amazon.com>