prov/sockets: retry setting conn->av_index
When we make dynamic connections, it is possible for a connect message
to arrive before the receiver issue fi_av_insert. This result in
conn->av_index being set to FI_ADDR_NOTAVAIL, and leading to future
message mismatch.

Retry setting the av_index fixes the issue. It works because the
matching will always happen after the receiver issues fi_av_insert or it
has to be a wild card receive.

Signed-off-by: Hui Zhou <hzhou321@anl.gov>