prov/cxi: Support auth_key hints
If users speciy an auth_key hint, this should be returned to the user
in the fi_getinfo fi_info out. The CXI provider was missing this
functionality.

When an auth_key is provided, an fi_info will only be returned to the
user if the auth_key is valid. Valid means there is a already configured
CXI service and the request VNI can be allocated against the CXI
service.

NETCASSINI-4317

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>