prov/net: Look-ahead in tcp stream
If we're waiting for an ACK or an RMA read response, it may be
blocked behind a unexpected receive message.  Unfortunately, this
is a common MPI flow:

    for each rank
        RMA read()
    barrier()

The barrier messages may begin to show up from peers before
the local rank has finished its RMA read operations.  And the
local rank won't post the receive message for the barrier until
the RMA read operations have completed.

There seems to be some sort of assumption that RMA read responses
have a separate from from messages, which is true when using an
RDMA device.  However, it's not true for tcp streams.  To handle
this case, we need to look ahead in the tcp stream for the response,
queuing the barrier message.

To minimize the changes to the provider and avoid re-introducing
unexpected message buffering at the receiver, we limit peeking
ahead in the tcp stream to a single 0-byte tagged message (what MPI
uses for barrier), at least for now.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>