Make fabtests capable of selecting an EFA device.
The --info-index option is used to specify the index of the desired
EFA device present in the cluster.
A test is also introduced in this commit that tests this functionality.
The test runs a fabtest using each EFA device present in the cluster
and checks for the transmitted bytes before and after the execution
of the fabtest.

Signed-off-by: Vishwas Dsouza <vidsouza@amazon.com>