prov/opx: fi_getInfo - Support multiple HFI devices
Previously when an application running on a server that supported multiple
HFI devices called the fi_getInfo interface, the OPX Provider returned
information for only the HFI_0 device.  Failure to return information for
all HFI devices caused applications like DAOS to not recognize OPX as a
provider supporting the other HFI devices.

Signed-off-by: Charles King <charles.king@cornelisnetworks.com>