prov/cxi: Support finding best auth_key
If the libfabric user fails to provide an auth_key as a hint (e.g.
fabtests), when the default service ID is disabled, these applications
will fail to run.

The solution is for an administrator to pre-configure an service ID for
such applications to use. The issue is these applications will not use
the service ID to generate an auth_key.

For these applications, the CXI provider will generate an auth_key on the
applications behalf. The CXI provider will attempt to generate an
auth_key using a service ID matching UID, then GID, and finally an
unrestricted service ID. If none of the above is found, the domain
auth_key will remain NULL. This will most likely result in fi_domain
allocation failures later.

NETCASSINI-4454

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>