prov/net: Expand saved rx support to include storing msg data
Currently, we only save and continue if the inbound message
contains no application data.  Expand this to allow storing a
small amount of application data (up to the inject size).

This will be expanded by a subsequent patch to allow saving
multiple messages, which is needed to handle MPI_Bsend call
semantics.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>