prov/efa: Remove HMEM check from rxr_ep_ctrl()
This function checked the domain capabilities for the FI_HMEM bit.
However, HMEM is a capability of a fabric endpoint, not a fabric domain
(see fi_domain(3), fi_endpoint(3)), so there is no circumstance where
FI_HMEM would ever be set in util_domain.info_domain_caps (at least when
this function was called).

Signed-off-by: Darryl Abbate <drl@amazon.com>