prov/linkx: Modify fi_info cache
Modify the access to the fi_info cache to match against domain name or
fabric name.

When we are setting the fi_fabric() we can grab the fi_info which matches
the fabric name. All domains which are part of the same fabric have the
same fabric information.

When we are setting the fi_domain(), we need to grab the fi_info which
matches the domain name. Each domain represents a different interface, and
we need to ensure that the domain requested by the application is honored.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>