prov/efa: Add p2p_required_by_impl to efa_hmem_info
This adds a p2p_required_by_impl field to the efa_hmem_info struct as a
way to track the minimum p2p requirements for each supported FI_HMEM
interface. Since the p2p requirements of each interface is more-or-less
unwritten/institutional knowledge, it should be defined in a single
place and queried as needed, instead of hardcoding in various places
throughout the codebase or deducing from a combination of other
parameters.

This also renames p2p_supported to p2p_supported_by_device to better
indicate the intention of the struct field.

Signed-off-by: Darryl Abbate <drl@amazon.com>