prov/efa: Implement FI_OPT_FI_HMEM_P2P validation
This adds a function for checking the validity of an endpoint p2p opt
against a given HMEM interface in a more generic way (i.e. without the
need to hardcode). The benefit is if/when the p2p requirements change
for any given HMEM interface, the p2p opt validation will work as
intended.

This also adds a p2p_disabled_by_user field to the efa_hmem_info struct,
which is initialized to false. The field is set to true when the user
calls fi_setopt() with FI_HMEM_P2P_DISABLED.

Signed-off-by: Darryl Abbate <drl@amazon.com>