core/hmem_cuda: adjust FI_HMEM_CUDA_ENABLE_XFER behavior
The environment variable FI_HMEM_CUDA_ENABLE_XFER control whether
CUDA API can be used to copy data. Currently, its help string states
that it default value is true, which means by default a provider
is safe to make CUDA calls.

However, by default it is not always safe for a provider to make
CUDA calls, and some providers does not make CUDA calls by default.
Such provider still need to make CUDA calls when user explictly
set this environment variable to true.

This patch used an enum (instead of a boolean) to store the
value of the user setting. The enum has three possible values:

    CUDA_XFER_DISABLED means the environment variable is set to 0
    CUDA_XFER_ENABLED means the environment variable is set to 1
    CUDA_XFER_UNSPECIED means the environment variable is not set

It then exposed the value of this setting through a public function,
such that provider can access.

It also updated the help string to:
"""
By default, each provider decide whether it will make CUDA calls.
"""
which better describes current situation.

It is worth noting that this patch did not change the current
behavior of any provider because:

Prior to this patch, the value of this environment variable is
not accessible to a provider.

Prior to this patch, if the envrionment variable was set to 0,
cuda_ops.cudaMemcpy was set to a no-op and cuda_ipc_enabled was
set to false. This patch maintained this behavior.

Signed-off-by: Wei Zhang <wzam@amazon.com>