fabtests/functional: Add data_validation option to fi_rdm_atomic
This change adds a mock for atomic operations, initializes values before
executing each atomic, and checks the final values for consistency with
the mocked operation.  This checking requires additional synchronization
which affects performance, so it is off by default.

The data_validation is enabled using the '-v' flag.  When the flag is
provided, a summary is printed and return value is EXIT_FAILURE if any
validations fail.

Signed-off-by: Luke Robison <lrbison@amazon.com>