prov/efa: use cuda_xfer_setting to decide whether to use gdrcopy
When cudaMemcpy is allowed, we want to use it and do not want
to use gdrcopy. Gdrcopy is used when the memory is registered
with gdrcopy.

Whether cudaMemcpy is allowed can now be determined by
calling cuda_get_xfer_setting(). This patch therefore
do not register memory with gdrcopy when cudaMemcpy is allowed.

Signed-off-by: Wei Zhang <wzam@amazon.com>