prov/efa: Move unit test teardown actions to teardown function
Previously, the teardown actions were part of the unit test code. With
this approach, any unrecoverable failure in one test would cause the
subsequent unit tests to fail.

This commit moves the teardown actions to a separate function that is
executed by the framework even if the unit test fails and allows for
subsequent unit tests to run successfully.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>