prov/efa: Abstract initial keygen values as macros
This keeps the initial values defined in a central location since
avoiding collisions with these values is critical.

This also adds a function to generate a key for an efa_mr's mr_fid.key
for FI_HMEM_CUDA when p2p is unavailable (to be used in a later commit).

Signed-off-by: Darryl Abbate <drl@amazon.com>