man/guide: Add user-guide to man pages
Relocate and update the ofi-guide from the adjacent git repo to
integrate with the libfabric man pages.  This will make it easier
to keep it in sync with API updates and avoid duplicate text
between the guide and other man pages.

Update man pages to align with guide, expanding descriptions
as needed, with appropriate references between man pages.
Remove duplicate and outdated text.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>