prov/efa: simplify implementation of DC longcts protocol
Prior to this patch, the TX entry of a DC longcts
protocol was released when both all DATA packets' send
completion have arrived and RECEIPT packet was released.

This is not ideal, because we want to write TX completion
as soon as RECEIPT was received.

The current implementation is because of two things:

First, there is no information in the packet
entry can be used to determine whether a DATA packet
is used by a DC longcts protocol or a normal longcts protocol,

Second, if a DATA packet is used by a normal longcts protocol,
the op_entry's byte_acked must be updated.

Beacuse of 1, we have to update tx_entry->byte_acked when we
receive the send completion of a DATA packet, which means
the tx_entry must not be released when we received the send
completion of the DATA packet.

As a result, we cannot release TX entry when we received
the RECEIPT packet.

This patch addressed the issue by introducing a flag

     RXR_PKT_ENTRY_DC_LONGCTS_DATA

and set it in the pkt_entry->flags for DATA packet used
by DC longcts protocol.

rxr_pkt_hand_data_send_completion() can check this flag
to determine whether a DATA packet is used by a DC longcts
protocol, and do not do anything for such a DATA packet.

The function rxr_pkt_handle_receipt_recv() was simplified
to just write TX completion.

The two tx_entry flags RXR_LONGCTS_PROTOCOL and RXR_RECEIPT_RECEIVED
were removed.

The two functions

    rxr_pkt_handle_dc_longcts_rtw_send_completion()
    rxr_pkt_handle_dc_longcts_rtm_send_completion()

were removed.

Signed-off-by: Wei Zhang <wzam@amazon.com>