ofi_net: Update ofi_bsock_send/ofi_bsock_sendv calls
Update the signature of the functions ofi_bsock_send and
ofi_bsock_sendv to match their recv/recvv counterparts.

The functions now return 0 on success and a negative value on
failure. The number of bytes consumed by the bsock interface for
the given call (whether the data is buffered in the bsock buffer
or sent on the socket) is returned in *len.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>