prov/opx: Cleanup pointers after free()
Make Opx do a better job of sanitizing pointers after freeing them

Many of these changes are just setting stack variables and not
changing underlying behavior, but it is still a good pattern to
have in place for future development.

Also fix a few use-after-free issues and some double-free issues
on close/teardown of Opx.

Signed-off-by: Tim Thompson <tim.thompson@cornelisnetworks.com>