prov/net: Fix memory leak found by ASAN
ASAN reports the following leak during normal executions:

Indirect leak of 72 byte(s) in 3 object(s) allocated from:
0 0xffffa828a4b4 in __interceptor_calloc ../../../../src/libsanitizer/asan/asan_malloc_linux.cpp:154
1 0xffffa7c6b600 in fid_list_insert prov/util/src/util_main.c:87
2 0xffffa7c65f58 in ofi_ep_bind_cq prov/util/src/util_ep.c:67
3 0xffffa7c66ee0 in ofi_ep_bind prov/util/src/util_ep.c:179
4 0xffffa7c7f93c in xnet_ep_bind prov/tcp/src/xnet_ep.c:601
5 0xffffa77fc198 in fi_ep_bind pkgbuild/libfabric/include/rdma/fi_endpoint.h:201

When binding the EP to the EQ, the EP's fid is added into a list
owned by the EQ with fid_list_insert(). This FID is never deleted
because the net provider doesn't call fid_list_remove() when
the EP is deleted.

This patch reworks the function xnet_ep_close() in order to call
ofi_endpoint_close(), which has the side-effect of fixing the
leak.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>