prov/opx: SDMA Completion workaround for driver cache invalidation race condition
This is a temporary workaround to address a data integrity problem caused by a
race condition in the HFI driver. OPX will need to wait for the driver to mark
an SDMA send request as completed before notifying the user that the send is
complete, rather than inferring that the send has completed after receiving ACKS
for all packets sent.

Signed-off-by: Ben Lynam <Ben.Lynam@cornelisnetworks.com>