prov/opx: Add reliability sanity check to detect when send buffer is illegally altered.
This change adds some additional sanity checking in debug builds when using
reliability replays that do not copy payload data into the replay's
bounce buffer. Up to the first 64 bytes of the payload data is copied into
the body of the replay when the replay is registered. Later on, when either
a NAK or an ACK is received for the packet, the copied data in the body of
the replay is then compared to the current value in the user's send buffer.
If there is any mismatch, an error message will be printed indicating that
the buffer was altered while the send was still in progress.

Signed-off-by: Ben Lynam <Ben.Lynam@cornelisnetworks.com>