util/av: Check for count = 0 in ofi_ip_av_insert
For some reason MPI wants to insert 0 addresses into the AV.
This is handled by ofi_ip_av_insertv(), but not directly by
ofi_ip_av_insert().  Prior to calling through to the insertv
call, it calls ofi_sizeofaddr(), which generates a warning of
an unknown address format.  This is because count is 0 and
addr parameter references randomness.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>