prov/tcp: Use context to mark message as claimed instead of tag bit
MPI uses the upper most bit for personal reasons (ignoring the
mem_tag_format).  As a result, the provider assumes that the tag
is for a claimed message and fails to match with posted
receives.  We can set the context for a claimed message instead
of toggling a tag bit.  This allows us to give the reserved bit
back to the app, which is using it anyway.

To avoid matching a claimed message we only match if the
saved context is still NULL.

This change assumes the app doesn't use NULL when claiming a
message, which is the case in practice and, really, the only
sensible thing to do.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>