prov/tcp: Fix counter update error for saved messages
When completing a saved message, we'll use the report_success
callback.  The xnet_report_success() call properly checks to
see if we're completing a saved message.  In which case, no
completion is actually written until the saved message is given
to the app (with corresponding matching receive).  However,
if the report_success callback references xnet_report_cntr_success(),
we still update the completion counter, even though the app hasn't
received the message.

To fix this use a common call to write completions to the cq
and remove the report_success callback.  This simplifies the code
and ensures that the proper checks are made prior to updating
either a counter or cq.  Also add a check for saved transfers
when reporting errors.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>