prov/efa: avoid unnecessary copy when sending data from shm
The function rxr_pkt_init_data_from_op_entry() prepare pkt_entry
to sent over wire. It has a fast path to send from user buffer
directly (without copy to bounce buffer).

Prior to this patch, it is not using fast path when shm is to
be used, and user did not provide memory descriptor, when
it can be used.

This patch addressed the issue

Signed-off-by: Wei Zhang <wzam@amazon.com>