prov/efa: move rxr_read_post_once() to rxr_pkt_entry_read()
This patch move rxr_read_post_once() to rxr_pkt_entry.c and
renamed it rxr_pkt_entry_read().

It also made one change to this function: rxr_read_post_once()
allocate a packet entry and use it as the context of the read
opreation. rxr_pkt_entry_read() does not allocate packet entry,
and expect caller to do so.

This is to prepare a future optimization that local read
operation do not allocate additional packet entry.

A local read operation copies data from a packet entry to
a receiver buffer. Currently, because rxr_read_post_once()
is used, we allocate an additonal packet entry and use
it as context of local read.

With this change, in future commit, we will use the packet
entry that contains data as context, and avoid allocating
additional packet entry.

Signed-off-by: Wei Zhang <wzam@amazon.com>