prov/tcp: Store cntr to update in xfer_entry instead of cntr_inc
Instead of storing a pointer to a function that increments a
counter, store the pointer to the counter to update.  This
allows the xnet_xfer_entry to directly reference which
counter, if any, needs updating, rather than accessing the
counter indirectly through the associated endpoint.

This change is needed to handle saved messages which complete
after the receiving endpoint has been destroyed.  It also
simplifies updating error counters, avoiding an unnecessary
lookup.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>